Changed mixSDL_Object and TFBSound_Object to be intptr_t for 64bit compatibility (bug #252)
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@1075 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/libs/sound/sound_chooser.h
@@ -39,7 +39,7 @@
  *  Interface Types
  */
 
-typedef unsigned int TFBSound_Object;
+typedef intptr_t TFBSound_Object;
 typedef int TFBSound_IntVal;
 typedef const int TFBSound_SourceProp;
 typedef const int TFBSound_BufferProp;
